Better chart coloring
Added a color randomizer for chart colors, which generates colors for
each pie section that is clearly visually distinct from the others. This
makes it easier to divide one section from another when looking at the
chart, and easier to identify elsewhere in the app when cgroups are
identified by color.

Also adds an error condition when the chart is set to contain more data
than it can contain in its 360 degree circle.